Remove pipeline flush from memmove and mempcpy
These calls no longer goes through the shared library linking system when
calling memcpy because this adds an overhead of several cycles caused by
following an function pointer (on top of any branch mispredict in memmove's
case). This has required that mempcpy is reimplemented in assembly so that
memcpy, memmove and mempcpy can be in the same source file.