Select implementation at runtime for memcpy, memmove and mempcpy
This relies upon the insertion of a trampoline function that loads the
address of the appropriate implementation from a static variable. The use of
a static initialiser that points at the function that patches up the function
pointer itself ensure that this is thread-safe.

The method used to determine which is the best implementation takes
advantage of the change in behaviour of interworking branches between ARMv6
and ARMv7. This makes the test OS-independent, and in any case it's
questionable how safe it is to be making system calls when a function as
fundamental as memcpy isn't functional yet, so this is probably a safer
approach than asking the kernel.

The main downside here is that anyone trying the library on an ARMv7 platform
which lacks NEON (the usual example is nVidia Tegra 2) will find it doesn't
work. But then we're not pretending to offer an optimised implementation for
that anyway, and they might as well continue to use glibc.