Limit the return values of ARMv7 memcmp()
When the input blocks are sized/aligned sufficiently to permit SIMD
comparison, the ARMv7 implementation of memcmp() was returning
INT32_MIN, 0 or INT32_MAX. This appears to cause compatibility issues with
some applications (presumably, the values are being narrowed to 16 bits so
INT32_MIN is interpreted as = 0 and INT32_MAX is interpreted as < 0).

While this implementation is technically conformant to the C standard, it
will be impractical to fix all the broken callers, so this patch ensures
that the return values in these cases are -1, 0 or +1 instead. Note that
for smaller and/or unaligned blocks, other values can still be returned,
but they will remain within the range [INT16_MIN, INT16_MAX] and so should
be less likely to cause probems.